+++ index.js
@@ -9,35 +9,30 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
 */
 
 'use strict';
-const hyd = require('hydrolysis');
 const dom5 = require('dom5');
-const Polymer = require('./lib/polymer-styling.js');
 
-const pathResolver = require('./lib/pathresolver');
-const initialValues = require('./lib/initial-values').initialValues;
+const pathResolver = require('./lib/pathresolver.js');
 
-/*
- * Apply Shim uses a dynamicly create DOM element
- * to lookup initial property values at runtime.
- *
- * Override with a fixed table for building
- */
-Polymer.ApplyShim._getInitialValueForProperty = (property) =>
-  initialValues[property] || '';
+let ApplyShim, CssParse, StyleTransformer, StyleUtil;
+const loadShadyCSS = require('./lib/shadycss-entrypoint.js');
 
+const path = require('path');
+
+/** @type {number} */
+const AnalyzerVersion = require(path.resolve(require.resolve('polymer-analyzer'), '../../package.json')).version.match(/\d+/)[0];
+
+const {Analyzer, InMemoryOverlayUrlLoader} = require('polymer-analyzer');
+
+// Use `Analysis` from polymer-analyzer for typing
+/* eslint-disable no-unused-vars */
+const {Analysis} = require('polymer-analyzer/lib/model/model.js');
+/* eslint-enable */
+
+const {dirShadowTransform, slottedToContent} = require('./lib/polymer-1-transforms.js');
 
 const pred = dom5.predicates;
 
 const domModuleCache = Object.create(null);
-
-const domModuleMatch = pred.AND(
-  pred.hasTagName('dom-module'),
-  pred.OR(
-    pred.hasAttr('id'),
-    pred.hasAttr('name'),
-    pred.hasAttr('is')
-  )
-);
 
 // TODO: upstream to dom5
 const styleMatch = pred.AND(
@@ -52,46 +47,27 @@ const styleMatch = pred.AND(
 
 const notStyleMatch = pred.NOT(styleMatch);
 
-const customStyleMatch = pred.AND(
-  pred.hasTagName('style'),
+const customStyleMatchV1 = pred.AND(
+  styleMatch,
   pred.hasAttrValue('is', 'custom-style')
+);
+
+const customStyleMatchV2 = pred.AND(
+  styleMatch,
+  pred.parentMatches(
+    pred.hasTagName('custom-style')
+  )
 );
 
 const styleIncludeMatch = pred.AND(styleMatch, pred.hasAttr('include'));
 
-// TODO: upstream to dom5
-const inlineScriptMatch = pred.AND(
-  pred.hasTagName('script'),
-  pred.OR(
-    pred.NOT(
-      pred.hasAttr('type')
-    ),
-    pred.hasAttrValue('type', 'text/javascript'),
-    pred.hasAttrValue('type', 'application/javascript')
-  ),
-  pred.NOT(
-    pred.hasAttr('src')
-  )
-);
-
 const scopeMap = new WeakMap();
-
-// TODO: upstream to dom5
-function ancestorWalk(node, match) {
-  while(node) {
-    if (match(node)) {
-      return node;
-    }
-    node = node.parentNode;
-  }
-  return null;
-}
 
 function prepend(parent, node) {
   if (parent.childNodes.length > 0) {
     dom5.insertBefore(parent, parent.childNodes[0], node);
   } else {
-    dom5.appendChild(parent, node);
+    dom5.append(parent, node);
   }
 }
 
@@ -99,30 +75,25 @@ function prepend(parent, node) {
  * Collect styles from dom-module
  * In addition, make sure those styles are inside a template
  */
-function getAndFixDomModuleStyles(module) {
+function getAndFixDomModuleStyles(domModule) {
   // TODO: support `.styleModules = ['module-id', ...]` ?
-  const styles = dom5.queryAll(module, styleMatch);
+  const styles = dom5.queryAll(domModule, styleMatch, undefined, dom5.childNodesIncludeTemplate);
   if (!styles.length) {
     return [];
   }
-  let template = dom5.query(module, pred.hasTagName('template'));
+  let template = dom5.query(domModule, pred.hasTagName('template'));
   if (!template) {
     template = dom5.constructors.element('template');
     const content = dom5.constructors.fragment();
     styles.forEach(s => dom5.append(content, s));
     dom5.append(template, content);
-    dom5.append(module, template);
+    dom5.append(domModule, template);
   } else {
-    styles.forEach(s => {
-      // create a template if one does not exist for this dom-module with styles
-      let templateContent = template.childNodes[0];
-      if (!templateContent) {
-        templateContent = dom5.constructors.fragment();
-        dom5.append(template, templateContent);
-      }
+    styles.forEach((s) => {
+      let templateContent = template.content;
       // ensure element styles are inside the template element
-      const parent = ancestorWalk(s, n =>
-        n === templateContent || n === module
+      const parent = dom5.nodeWalkAncestors(s, (n) =>
+        n === templateContent || n === domModule
       );
       if (parent !== templateContent) {
         prepend(templateContent, s);
@@ -152,19 +123,19 @@ function inlineStyleIncludes(style) {
   const includes = getAttributeArray(style, 'include');
   const leftover = [];
   const baseDocument = style.__ownerDocument;
-  includes.forEach(id => {
-    const module = domModuleCache[id];
-    if (!module) {
+  includes.forEach((id) => {
+    const domModule = domModuleCache[id];
+    if (!domModule) {
       // we missed this one, put it back on later
       leftover.push(id);
       return;
     }
-    const includedStyles = getAndFixDomModuleStyles(module);
+    const includedStyles = getAndFixDomModuleStyles(domModule);
     // gather included styles
-    includedStyles.forEach(ism => {
+    includedStyles.forEach((ism) => {
       // this style may also have includes
       inlineStyleIncludes(ism);
-      const inlineDocument = module.__ownerDocument;
+      const inlineDocument = domModule.__ownerDocument;
       let includeText = dom5.getTextContent(ism);
       // adjust paths
       includeText = pathResolver.rewriteURL(inlineDocument, baseDocument, includeText);
@@ -192,45 +163,27 @@ function applyShim(ast) {
    * Decorated <style> elements are ones with `__cssRules` property
    * with a value of the CSS ast
    */
-  Polymer.StyleUtil.forEachRule(ast, rule => Polymer.ApplyShim.transformRule(rule));
+  StyleUtil.forEachRule(ast, (rule) => ApplyShim.transformRule(rule));
 }
 
 function getModuleDefinition(moduleName, elementDescriptors) {
   for (let ed of elementDescriptors) {
-    if (ed.is.toLowerCase() === moduleName) {
+    if (ed.tagName && ed.tagName.toLowerCase() === moduleName) {
       return ed;
     }
   }
   return null;
 }
 
-function getTypeExtends(elementDescriptor) {
-  let props = elementDescriptor.properties || [];
-  let ret = '';
-  // loop over element properties with javascript AST
-  for (let i = 0; i < props.length; i++) {
-    if (props[i].name === 'extends') {
-      // node is espree AST node
-      let node = props[i].javascriptNode;
-      // property node has a value node with key and value nodes
-      if (node && node.type === 'Property') {
-        ret = node.value.value;
-        break;
-      }
-    }
-  }
-  return ret.toLowerCase();
-}
-
-function shadyShim(ast, style, elements) {
+function shadyShim(ast, style, analysis) {
   const scope = scopeMap.get(style);
-  const moduleDefinition = getModuleDefinition(scope, elements);
+  const moduleDefinition = getModuleDefinition(scope, analysis.getFeatures({kind: 'element'}));
   // only shim if module is a full polymer element, not just a style module
   if (!scope || !moduleDefinition) {
     return;
   }
-  const ext = getTypeExtends(moduleDefinition);
-  Polymer.StyleTransformer.css(ast, scope, ext);
+  const ext = moduleDefinition.extends;
+  StyleTransformer.css(ast, scope, ext);
 }
 
 function addClass(node, className) {
@@ -244,146 +197,212 @@ function addClass(node, className) {
   dom5.setAttribute(node, 'class', classList.join(' '));
 }
 
-function markElement(domModule, scope) {
-  const buildType = Polymer.Settings.useNativeShadow ? 'shadow' : 'shady';
+function markElement(domModule, scope, useNativeShadow) {
+  const buildType = useNativeShadow ? 'shadow' : 'shady';
+  // apply scoping to dom-module
   dom5.setAttribute(domModule, 'css-build', buildType);
-  // mark elements' subtree under shady build
-  if (buildType === 'shady' && scope) {
-    const template = dom5.query(domModule, pred.hasTagName('template'));
-    // apply scoping to template
-    if (template) {
-      const elements = dom5.queryAll(template, notStyleMatch);
-      elements.forEach(el => addClass(el, scope));
+  // apply scoping to template
+  const template = dom5.query(domModule, pred.hasTagName('template'));
+  if (template) {
+    dom5.setAttribute(template, 'css-build', buildType);
+    // mark elements' subtree under shady build
+    if (buildType === 'shady' && scope) {
+      const elements = dom5.queryAll(template, notStyleMatch, undefined, dom5.childNodesIncludeTemplate);
+      elements.forEach((el) => addClass(el, scope));
     }
   }
 }
 
-function slotToContent(ast) {
-  Polymer.StyleUtil.forEachRule(ast, (rule) => {
-    rule.selector = Polymer.StyleTransformer._slottedToContent(rule.selector);
+// For forward compatibility with ShadowDOM v1 and Polymer 2.x,
+// replace ::slotted(${inner}) with ::content > ${inner}
+function slottedTransform(ast) {
+  StyleUtil.forEachRule(ast, (rule) => {
+    rule.selector = slottedToContent(rule.selector);
   });
 }
 
 function dirTransform(ast) {
-  Polymer.StyleUtil.forEachRule(ast, (rule) => {
-    rule.selector = Polymer.StyleTransformer._dirShadowTransform(rule.selector);
+  StyleUtil.forEachRule(ast, (rule) => {
+    rule.selector = dirShadowTransform(rule.selector);
   });
 }
 
-function polymerCssBuild(paths, options) {
-  if (options && options['build-for-shady']) {
-    Polymer.Settings.useNativeShadow = false;
+function setUpLibraries(useNativeShadow) {
+  ({
+    ApplyShim,
+    CssParse,
+    StyleTransformer,
+    StyleUtil
+  } = loadShadyCSS(useNativeShadow));
+}
+
+function setNodeFileLocation(node, analysisKind) {
+  if (!node.__ownerDocument) {
+    node.__ownerDocument = analysisKind.sourceRange.file;
   }
-  const nativeShadow = Polymer.Settings.useNativeShadow;
-  // build hydrolysis loader
-  const loader = new hyd.Loader();
-  // ignore all files we can't find
-  loader.addResolver(new hyd.NoopResolver({test: () => true}));
+}
+
+/**
+ *
+ * @param {Analysis} analysis
+ * @param {Function} query
+ * @param {Function=} queryOptions
+ * @return {Array}
+ */
+function nodeWalkAllDocuments(analysis, query, queryOptions = undefined) {
+  const results = [];
+  for (const document of analysis.getFeatures({kind: 'html-document'})) {
+    const matches = dom5.nodeWalkAll(document.parsedDocument.ast, query, undefined, queryOptions);
+    matches.forEach((match) => {
+      setNodeFileLocation(match, document);
+    });
+    results.push(...matches);
+  }
+  return results;
+}
+
+/**
+ * Handle the difference between analyzer 2 and 3 Analyzer::getDocument
+ * @param {!Analysis} analysis
+ * @param {string} url
+ * @return {!Document}
+ */
+function getDocument(analysis, url) {
+  const res = analysis.getDocument(url);
+  if (res.error) {
+    throw res.error;
+  }
+  if (res.value) {
+    return res.value;
+  }
+  return res;
+}
+
+function getAstNode(domModule) {
+  if (AnalyzerVersion === '2') {
+    return domModule.astNode
+  } else {
+    return domModule.astNode.node;
+  }
+}
+
+function getOrderedDomModules(analysis) {
+  const domModules = [];
+  for (const document of analysis.getFeatures({kind: 'html-document'})) {
+    for (const domModule of document.getFeatures({kind: 'dom-module'})) {
+      domModules.push(domModule)
+    }
+  }
+  return domModules;
+}
+
+async function polymerCssBuild(paths, options = {}) {
+  const nativeShadow = options ? !options['build-for-shady'] : true;
+  const polymerVersion = options['polymer-version'] || 2;
+  const customStyleMatch = polymerVersion === 2 ? customStyleMatchV2 : customStyleMatchV1;
+  setUpLibraries(nativeShadow);
+  // build analyzer loader
+  const loader = new InMemoryOverlayUrlLoader();
+  const analyzer = new Analyzer({urlLoader: loader});
   // load given files as strings
-  paths.forEach(p => {
-    loader.addResolver(new hyd.StringResolver(p));
+  paths.forEach((p) => {
+    loader.urlContentsMap.set(analyzer.resolveUrl(p.url), p.content);
   });
-  const analyzer = new hyd.Analyzer(true, loader);
   // run analyzer on all given files
-  return Promise.all(
-    paths.map(p => analyzer.metadataTree(p.url))
-  ).then(() => {
-    // un-inline scripts that hydrolysis accendentally inlined
-    analyzer.nodeWalkAllDocuments(inlineScriptMatch).forEach(script => {
-      if (script.__hydrolysisInlined) {
-        dom5.setAttribute(script, 'src', script.__hydrolysisInlined);
-        dom5.setTextContent(script, '');
-      }
+  /** @type {Analysis} */
+  const analysis = await analyzer.analyze(paths.map((p) => p.url));
+  // map dom modules to styles
+  const moduleStyles = [];
+  for (const domModule of getOrderedDomModules(analysis)) {
+    const id = domModule.id;
+    const scope = id.toLowerCase();
+    const el = getAstNode(domModule);
+    domModuleCache[scope] = el;
+    setNodeFileLocation(el, domModule);
+    markElement(el, scope, nativeShadow);
+    const styles = getAndFixDomModuleStyles(el);
+    styles.forEach((s) => {
+      scopeMap.set(s, scope);
+      setNodeFileLocation(s, domModule);
     });
-  }).then(() => {
-    // map dom modules to styles
-    return analyzer.nodeWalkAllDocuments(domModuleMatch).map(el => {
-      const id = dom5.getAttribute(el, 'id') || dom5.getAttribute(el, 'name') ||
-          dom5.getAttribute(el, 'is');
-      if (!id) {
-        return [];
+    moduleStyles.push(styles);
+  }
+  // inline and flatten styles into a single list
+  const flatStyles = [];
+  moduleStyles.forEach((styles) => {
+    if (!styles.length) {
+      return;
+    }
+    // do style includes
+    if (options ? !options['no-inline-includes'] : true) {
+      styles.forEach((s) => inlineStyleIncludes(s));
+    }
+    // reduce styles to one
+    const finalStyle = styles[styles.length - 1];
+    dom5.setAttribute(finalStyle, 'scope', scopeMap.get(finalStyle));
+    if (styles.length > 1) {
+      const consumed = styles.slice(0, -1);
+      const text = styles.map((s) => dom5.getTextContent(s));
+      const includes = styles.map((s) => getAttributeArray(s, 'include')).reduce((acc, inc) => acc.concat(inc));
+      consumed.forEach((c) => dom5.remove(c));
+      dom5.setTextContent(finalStyle, text.join(''));
+      const oldInclude = getAttributeArray(finalStyle, 'include');
+      const newInclude = oldInclude.concat(includes).join(' ');
+      if (newInclude) {
+        dom5.setAttribute(finalStyle, 'include', newInclude);
       }
-      const scope = id.toLowerCase();
-      // populate cache
-      domModuleCache[scope] = el;
-      // mark the module as built
-      markElement(el, scope);
-      const styles = getAndFixDomModuleStyles(el);
-      styles.forEach(s => scopeMap.set(s, scope));
-      return styles;
-    });
-  }).then(moduleStyles => {
-    // inline and flatten styles into a single list
-    const flatStyles = [];
-    moduleStyles.forEach(styles => {
-      if (!styles.length) {
-        return;
-      }
-      // do style includes
-      if (options && !options['no-inline-includes']) {
-        styles.forEach(s => inlineStyleIncludes(s));
-      }
-      // reduce styles to one
-      const finalStyle = styles[styles.length - 1];
-      dom5.setAttribute(finalStyle, 'scope', scopeMap.get(finalStyle));
-      if (styles.length > 1) {
-        const consumed = styles.slice(0, -1);
-        const text = styles.map(s => dom5.getTextContent(s));
-        const includes = styles.map(s => getAttributeArray(s, 'include')).reduce((acc, inc) => acc.concat(inc));
-        consumed.forEach(c => dom5.remove(c));
-        dom5.setTextContent(finalStyle, text.join(''));
-        const oldInclude = getAttributeArray(finalStyle, 'include');
-        const newInclude = oldInclude.concat(includes).join(' ');
-        if (newInclude) {
-          dom5.setAttribute(finalStyle, 'include', newInclude);
+    }
+    flatStyles.push(finalStyle);
+  });
+  // find custom styles
+  const customStyles = nodeWalkAllDocuments(analysis, customStyleMatch);
+  // inline custom styles with includes
+  if (options ? !options['no-inline-includes'] : true) {
+    customStyles.forEach((s) => inlineStyleIncludes(s));
+  }
+  // add custom styles to the front
+  // custom styles may define mixins for the whole tree
+  flatStyles.unshift(...customStyles);
+  // populate mixin map
+  flatStyles.forEach((s) => {
+    const text = dom5.getTextContent(s);
+    const ast = CssParse.parse(text);
+    applyShim(ast);
+  });
+  // parse, transform, emit
+  flatStyles.forEach((s) => {
+    let text = dom5.getTextContent(s);
+    const ast = CssParse.parse(text);
+    if (customStyleMatch(s)) {
+      // custom-style `:root` selectors need to be processed to `html`
+      StyleUtil.forEachRule(ast, (rule) => {
+        if (options && options['build-for-shady']) {
+          StyleTransformer.documentRule(rule);
+        } else {
+          StyleTransformer.normalizeRootSelector(rule);
         }
-      }
-      flatStyles.push(finalStyle);
-    });
-    return flatStyles;
-  }).then(styles => {
-    // find custom styles
-    const customStyles = analyzer.nodeWalkAllDocuments(customStyleMatch);
-    // inline custom styles with includes
-    customStyles.forEach(s => inlineStyleIncludes(s));
-    // add in custom styles
-    return customStyles.concat(styles);
-  }).then(styles => {
-    // populate mixin map
-    styles.forEach(s => {
-      const text = dom5.getTextContent(s);
-      const ast = Polymer.CssParse.parse(text);
-      applyShim(ast);
-    });
-    // parse, transform, emit
-    styles.forEach(s => {
-      let text = dom5.getTextContent(s);
-      const ast = Polymer.CssParse.parse(text);
-      if (customStyleMatch(s)) {
-        // custom-style `:root` selectors need to be processed to `html`
-        Polymer.StyleUtil.forEachRule(ast, rule => {
-          Polymer.StyleTransformer.documentRule(rule);
-        });
-        // mark the style as built
-        markElement(s);
-      }
-      applyShim(ast);
-      if (nativeShadow) {
-        slotToContent(ast);
+      });
+      // mark the style as built
+      markElement(s, null, nativeShadow);
+    }
+    applyShim(ast);
+    if (nativeShadow) {
+      if (polymerVersion === 1) {
+        slottedTransform(ast);
         dirTransform(ast);
-      } else {
-        shadyShim(ast, s, analyzer.elements);
       }
-      text = Polymer.CssParse.stringify(ast, true);
-      dom5.setTextContent(s, text);
-    });
-  }).then(() => {
-    return paths.map(p => {
-      return {
-        url: p.url,
-        content: dom5.serialize(analyzer.parsedDocuments[p.url])
-      };
-    });
+    } else {
+      shadyShim(ast, s, analysis);
+    }
+    text = CssParse.stringify(ast, true);
+    dom5.setTextContent(s, text);
+  });
+  return paths.map((p) => {
+    const doc = getDocument(analysis, p.url);
+    return {
+      url: p.url,
+      content: doc.parsedDocument.stringify()
+    };
   });
 }